requirements.txt
asgiref==3.7.2
Django==4.2.7
sqlparse==0.4.4
Pillow==10.1.0
flower==2.0.1
celery==5.3.4
stripe==7.6.0
WeasyPrint==60.1
redis==5.0.1
psycopg2-binary==2.9.9
django-constance[redis]==3.1.0
flake8==6.1.0
pre-commit==3.5.0
black==23.11.0
isort==5.12.0
django-environ==0.11.2
django-debug-toolbar==4.2.0
sentry-sdk==1.32.0
django-redis==5.4.0
gunicorn==21.2.0
django-extensions==3.2.3
django-cors-headers==4.3.0
python-dotenv==1.0.0
whitenoise==6.6.0